+++ srcs/sprite.c
@@ -15,14 +15,14 @@ void draw_sprites(t_struct *data, float x_off, float y_off, int index)
 	size = data->sprite[index].size;
 	while (i++ < size)
 	{
-		if (x_off + i < 0 || x_off + i > data->w_width)
+		if (x_off + i < 0 || x_off + i > (float )data->w_width)
 			continue;
 		if (g_rays[(int)(x_off + i)].distance <= data->sprite[index].dis)
 			continue;
 		j = 0;
 		while (j++ < size)
 		{
-			if (y_off + j < 0 || y_off + j > data->w_height)
+			if (y_off + j < 0 || y_off + j > (float )data->w_height)
 				continue;
 			id = SPRITE_SIZE * (SPRITE_SIZE * j / size) + (SPRITE_SIZE * i / size);
 			id = id >= (e) ? (e - 1) : id;
@@ -53,17 +53,17 @@ void set_up_sprite(t_struct *data)
 	{
 		data->sprite[i_spt].dis = distance_between_points(data->x, data->y,
 														  data->sprite[i_spt].x, data->sprite[i_spt].y);
-		spt_angle = atan2(data->sprite[i_spt].y - data->y, data->sprite[i_spt].x - data->x);
+		spt_angle = atan2f(data->sprite[i_spt].y - data->y, data->sprite[i_spt].x - data->x);
 		while (spt_angle - data->rotation_angle > M_PI)
 			spt_angle -= 2 * M_PI;
 		while (spt_angle - data->rotation_angle < -M_PI)
 			spt_angle += 2 * M_PI;
 		if (data->w_height > data->w_width)
-			data->sprite[i_spt].size = (data->w_height / data->sprite[i_spt].dis) * SQUARE_SIZE;
+			data->sprite[i_spt].size = ((float )data->w_height / data->sprite[i_spt].dis) * SQUARE_SIZE;
 		else
-			data->sprite[i_spt].size = (data->w_width / data->sprite[i_spt].dis) * SQUARE_SIZE;
-		data->sprite[i_spt].y_off = data->w_height / 2 - data->sprite[i_spt].size / 2;
-		data->sprite[i_spt].x_off = ((DEG(spt_angle) - DEG(data->rotation_angle)) * data->w_width) / SPRITE_SIZE + ((data->w_width / 2) - (data->sprite[i_spt].size / 2));
+			data->sprite[i_spt].size = ((float )data->w_width / data->sprite[i_spt].dis) * SQUARE_SIZE;
+		data->sprite[i_spt].y_off = (float )data->w_height / 2 - data->sprite[i_spt].size / 2;
+		data->sprite[i_spt].x_off = ((DEG(spt_angle) - DEG(data->rotation_angle)) * data->w_width) / SPRITE_SIZE + (((float )data->w_width / 2) - (data->sprite[i_spt].size / 2));
 		draw_sprites(data, data->sprite[i_spt].x_off, data->sprite[i_spt].y_off, i_spt);
 		i_spt++;
 	}
@@ -89,8 +89,8 @@ void initialize_sprite(t_struct *data)
 		{
 			if (data->map[i][j] == '2')
 			{
-				data->sprite[i_spt].x = (j + 0.5f) * SQUARE_SIZE;
-				data->sprite[i_spt].y = (i + 0.5f) * SQUARE_SIZE;
+				data->sprite[i_spt].x = ((float )j + 0.5f) * SQUARE_SIZE;
+				data->sprite[i_spt].y = ((float )i + 0.5f) * SQUARE_SIZE;
 				i_spt++;
 			}
 			j++;
